Refine missing import quick fix implementation
